Refactor the whole application
The application felt very un-professional so I will begin again from the
beginning. The changes in this commit remove the old code and start by
creating a get package and then listing all repositories that contain
the terms "cloud-platform-terraform" in the name. This avoids the need
to obtain the repository names before hand.